Moved Contact component to page folder
