+++ pages/contact/index.js
@@ -1,4 +1,4 @@
-import Contact from "components/Contact";
+import Contact from "./Contact";
 
 const ContactPage = () => {
   return (
